% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_boot_distns.R
\name{adjust_method}
\alias{adjust_method}
\title{Helper function to adjust the method in the DataFrame to help us
distinguish when plotting}
\usage{
adjust_method(method, group_algo)
}
\arguments{
\item{method}{The method for the particular experiment}

\item{group_algo}{The grouping algorithm used for the experiment}
}
\value{
Adjusted DataFrame w/ new methods
}
\description{
Helper function to adjust the method in the DataFrame to help us
distinguish when plotting
}
\keyword{internal}
