% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_boot_distns.R
\name{adjust_title}
\alias{adjust_title}
\title{Helper function to map the basename of the working directory to an
abbreviation for the plot title}
\usage{
adjust_title(basename)
}
\arguments{
\item{basename}{Basename of the working directory}
}
\value{
String to use for the plot title
}
\description{
Helper function to map the basename of the working directory to an
abbreviation for the plot title
}
