% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tex_code.R
\name{build_exp_diff}
\alias{build_exp_diff}
\title{Helper function to build the experiment difference DataFrame}
\usage{
build_exp_diff(pair_df, exp_df)
}
\arguments{
\item{pair_df}{DataFrame defining the unique experiment pairs}

\item{exp_df}{DataFrame detailing all experiment settings}
}
\value{
Experiment pair DataFrame
}
\description{
Helper function to build the experiment difference DataFrame
}
\keyword{internal}
