% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{build_legend_labeller}
\alias{build_legend_labeller}
\title{Helper function to create a legend labeller}
\usage{
build_legend_labeller(distinguisher)
}
\arguments{
\item{distinguisher}{Distinguishing factor for the experiment}
}
\value{
Named vector to map the legend labels
}
\description{
Helper function to create a legend labeller
}
\keyword{internal}
