% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_top1_plot.R
\name{build_nt1}
\alias{build_nt1}
\title{Creates the Node Top 1 plot to compare the various hierarchical classifiers}
\usage{
build_nt1(wd, ...)
}
\arguments{
\item{wd}{Working directory to find the data}

\item{...}{Additional arguments that can be passed for filtering}
}
\value{
Nothing; saves the plot to disk
}
\description{
Creates the Node Top 1 plot to compare the various hierarchical classifiers
}
