% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{check_valid_combo}
\alias{check_valid_combo}
\title{Helper function to determine if a combination of indices belong together}
\usage{
check_valid_combo(exp_diff_df, idx)
}
\arguments{
\item{exp_diff_df}{Experiment difference DataFrame}

\item{idx}{Indices that are being checked for the combination}
}
\value{
Boolean of whether these indices belong together
}
\description{
Helper function to determine if a combination of indices belong together
}
\keyword{internal}
