% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_train_time.R
\name{combine_cols}
\alias{combine_cols}
\title{Helper function which combines the metric.x/y and value.x/y in the
DataFrame}
\usage{
combine_cols(df)
}
\arguments{
\item{df}{DataFrame containing the multiple metric and value columns}
}
\value{
DataFrame with a single metric and value column
}
\description{
Helper function which combines the metric.x/y and value.x/y in the
DataFrame
}
\keyword{internal}
