% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{combine_like_columns}
\alias{combine_like_columns}
\title{Helper function to deal with columns like setting1, setting2 and
convert them into a single vector}
\usage{
combine_like_columns(df, identifier)
}
\arguments{
\item{df}{DataFrame containing the columns}

\item{identifier}{Identifier that distinguishes the column}
}
\value{
Vector which flattens all the entries
}
\description{
Helper function to deal with columns like setting1, setting2 and
convert them into a single vector
}
\keyword{internal}
