% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_train_time.R
\name{compute_total_time}
\alias{compute_total_time}
\title{Helper function to compute the total training time for a the HC DataFrame}
\usage{
compute_total_time(df)
}
\arguments{
\item{df}{HC time results DataFrame}
}
\value{
DataFrame containing only the total training time for a given
experiment
}
\description{
Helper function to compute the total training time for a the HC DataFrame
}
\keyword{internal}
