% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{convert_group_algo_name}
\alias{convert_group_algo_name}
\title{Converts the grouping algorithm names}
\usage{
convert_group_algo_name(setting)
}
\arguments{
\item{setting}{The value for the experiment}
}
\value{
More readable string for the grouping algorithm
}
\description{
Converts the grouping algorithm names
}
\keyword{internal}
