% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{convert_method_names}
\alias{convert_method_names}
\title{Converts the method names into a more read-able format for the plot title}
\usage{
convert_method_names(setting)
}
\arguments{
\item{setting}{The particular setting for the given experiment}
}
\value{
More read-able string for the given method
}
\description{
Converts the method names into a more read-able format for the plot title
}
\keyword{internal}
