% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{count_uniq_vals}
\alias{count_uniq_vals}
\title{Helper function to count the number of unique entries in the specified
columns of a DataFrame}
\usage{
count_uniq_vals(df, cols, idx)
}
\arguments{
\item{df}{DataFrame to count the unique entries}

\item{cols}{Columns to use for counting}

\item{idx}{Indices to use for counting}
}
\value{
Integer vector containing the number of unique entries in each column
}
\description{
Helper function to count the number of unique entries in the specified
columns of a DataFrame
}
\keyword{internal}
