% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_boot_distns.R
\name{filter_df}
\alias{filter_df}
\title{Helper function to do additional filtering of a DataFrame based on ...
argments}
\usage{
filter_df(df, addl_args)
}
\arguments{
\item{df}{The DataFrame to filter}

\item{addl_args}{Additional arguments passed to the plot function call}
}
\value{
Updated DataFrame
}
\description{
Helper function to do additional filtering of a DataFrame based on ...
argments
}
