% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_boot_distns.R
\name{gen_boot_distn}
\alias{gen_boot_distn}
\title{Produces the bootstrap distribution plot for a series of experiments}
\usage{
gen_boot_distn(wd, ...)
}
\arguments{
\item{wd}{Working directory to find the data}

\item{...}{Additional arguments that can be passed for filtering}
}
\value{
Nothing; saves the plot to disk
}
\description{
Produces the bootstrap distribution plot for a series of experiments
}
