% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{gen_boot_distns}
\alias{gen_boot_distns}
\title{Creates and saves the bootstrap distributions to disk}
\usage{
gen_boot_distns(pair_combo, update_df, boot_df, exp_diff_df, pair_df,
  exp_df, savepath)
}
\arguments{
\item{pair_combo}{Unique experiment pair ID}

\item{update_df}{DataFrame which could add new data to help extend the plot}

\item{boot_df}{Bootstrap results DataFrame}

\item{exp_diff_df}{Experiment difference DataFrame}

\item{pair_df}{DataFrame defining the experiment pairs}

\item{exp_df}{DataFrame detailing all experiment settings}

\item{savepath}{Directory in which to save the plot}
}
\value{
Nothing; saves the plot to disk
}
\description{
Creates and saves the bootstrap distributions to disk
}
\keyword{internal}
