% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{gen_boot_hists}
\alias{gen_boot_hists}
\title{Creates and saves the bootstrap histograms to disk}
\usage{
gen_boot_hists(pair_id, boot_df, exp_diff_df, savepath)
}
\arguments{
\item{pair_id}{Unique experiment pair ID}

\item{boot_df}{Bootstrap results DataFrame}

\item{exp_diff_df}{Experiment difference DataFrame}

\item{savepath}{Directory in which to save the plot}
}
\value{
Nothing; saves the plot to disk
}
\description{
Creates and saves the bootstrap histograms to disk
}
