% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tex_code.R
\name{gen_boot_res_tex}
\alias{gen_boot_res_tex}
\title{Generate the TeX code for the bootstrap results}
\usage{
gen_boot_res_tex(boot_df)
}
\arguments{
\item{boot_df}{Bootstrap results DataFrame}
}
\value{
LaTeX code which can be saved to disk
}
\description{
Generate the TeX code for the bootstrap results
}
