% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{gen_combo_distn}
\alias{gen_combo_distn}
\title{Function to generate a plot under the condition that there are two
distinguishing factors for an experiment (ex: method and estimator)}
\usage{
gen_combo_distn(boot_df, info_df, plot_title)
}
\arguments{
\item{boot_df}{Bootstrap results DataFrame}

\item{info_df}{DataFrame containing the plotting info for the \code{boot_df}}

\item{plot_title}{The plot title}
}
\value{
ggplot for bootstrap distributions
}
\description{
Function to generate a plot under the condition that there are two
distinguishing factors for an experiment (ex: method and estimator)
}
\keyword{internal}
