% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{gen_combo_hist}
\alias{gen_combo_hist}
\title{Function to generate a plot under the condition that there are two
distinguishing factors for an experiment (ex: method and estimator)}
\usage{
gen_combo_hist(pair_combo, boot_df, update_df, figure_titles)
}
\arguments{
\item{pair_combo}{Combination of valid pair_ids}

\item{boot_df}{Bootstrap results DataFrame}

\item{update_df}{DataFrame which adds the information we need to extend
the plot}

\item{figure_titles}{List containing the plot and caption title}
}
\value{
ggplot for bootstrap histograms
}
\description{
Function to generate a plot under the condition that there are two
distinguishing factors for an experiment (ex: method and estimator)
}
\keyword{internal}
