% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tex_code.R
\name{gen_exp_setting_tex}
\alias{gen_exp_setting_tex}
\title{Generate the experiment difference TeX code}
\usage{
gen_exp_setting_tex(pair_df, exp_df)
}
\arguments{
\item{pair_df}{DataFrame defining the unique experiment pairs}

\item{exp_df}{DataFrame detailing all experiment settings}
}
\value{
LaTeX code defining the unique experiments
}
\description{
Generate the experiment difference TeX code
}
