% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{gen_facet_names}
\alias{gen_facet_names}
\title{Helper function generate the facet names}
\usage{
gen_facet_names()
}
\value{
Named vector of metrics
}
\description{
Helper function generate the facet names
}
\keyword{internal}
