% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_stability.R
\name{gen_leaf_plot}
\alias{gen_leaf_plot}
\title{Generates the leaf entropy plot}
\usage{
gen_leaf_plot(wd, ...)
}
\arguments{
\item{wd}{Working directory to find the data}

\item{...}{Additional arguments that can be passed for filtering}
}
\value{
Nothing; saves the plot to disk
}
\description{
Generates the leaf entropy plot
}
