% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{gen_lone_hist}
\alias{gen_lone_hist}
\title{Function to generate a plot under the condition that there is only one
pair_combo that is valid}
\usage{
gen_lone_hist(pair_combo, boot_df, figure_titles)
}
\arguments{
\item{pair_combo}{Lone pair_id}

\item{boot_df}{Bootstrap results DataFrame}

\item{figure_titles}{List containing the plot and caption title}
}
\value{
ggplot for bootstrap histograms
}
\description{
Function to generate a plot under the condition that there is only one
pair_combo that is valid
}
\keyword{internal}
