% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{gen_plot_titles}
\alias{gen_plot_titles}
\title{Helper function to generate the plot and caption title based off the
experiment distinguishing factor}
\usage{
gen_plot_titles(exp_diff_df)
}
\arguments{
\item{exp_diff_df}{Experiment difference DataFrame}
}
\value{
List containing the plot and caption title
}
\description{
Helper function to generate the plot and caption title based off the
experiment distinguishing factor
}
\keyword{internal}
