% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tex_code.R
\name{gen_tex_code}
\alias{gen_tex_code}
\title{Generate all of the LaTeX code we will use to report the results from the
bootstrap experiments}
\usage{
gen_tex_code(wd)
}
\arguments{
\item{wd}{Directory containing the files needed to generate the TeX code}
}
\value{
Nothing; saves the .txt files to disk
}
\description{
Generate all of the LaTeX code we will use to report the results from the
bootstrap experiments
}
