% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_train_time.R
\name{gen_time_plot}
\alias{gen_time_plot}
\title{Generates the time comparison plot among the various methods}
\usage{
gen_time_plot(wd, metric = "leaf_top1")
}
\arguments{
\item{wd}{Working directory to find the data}

\item{metric}{Which metric to use for the time comparison plot}
}
\value{
Nothing; saves the plot to disk
}
\description{
Generates the time comparison plot among the various methods
}
