% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tex_code.R
\name{get_avg_diff}
\alias{get_avg_diff}
\title{Computes the average difference for a given metric (ex: leaf_top1)}
\usage{
get_avg_diff(boot_df, id, metric)
}
\arguments{
\item{boot_df}{Bootstrap results DataFrame}

\item{id}{Unique experiment pair ID}

\item{metric}{Metric for the particular experiment (ex: leaf_top1)}
}
\value{
DataFrame containing the ID the metric, and the mean difference
}
\description{
Computes the average difference for a given metric (ex: leaf_top1)
}
\keyword{internal}
