% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tex_code.R
\name{get_boot_table}
\alias{get_boot_table}
\title{Gets the bootstrap results for all the metrics}
\usage{
get_boot_table(boot_df)
}
\arguments{
\item{boot_df}{Bootstrap results DataFrame}
}
\value{
A wide DataFrame containing the average difference for all
experiments
}
\description{
Gets the bootstrap results for all the metrics
}
