% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tex_code.R
\name{get_exp_diff}
\alias{get_exp_diff}
\title{Infers the unique experiments that we will use to generate the TeX code
and the plots}
\usage{
get_exp_diff(pair_id, pair_df, exp_df)
}
\arguments{
\item{pair_id}{Unique experiment pair ID}

\item{pair_df}{DataFrame defining the unique experiment pairs}

\item{exp_df}{DataFrame detailing all experiment settings}
}
\value{
DataFrame containing the unique experiments and what distinguishes
them
}
\description{
Infers the unique experiments that we will use to generate the TeX code
and the plots
}
