% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{get_experiment_settings}
\alias{get_experiment_settings}
\title{Helper function to get the setting values for a given experiment}
\usage{
get_experiment_settings(exp_diff_df)
}
\arguments{
\item{exp_diff_df}{Experiment difference DataFrame}
}
\value{
Character vector of experiment settings
}
\description{
Helper function to get the setting values for a given experiment
}
\keyword{internal}
