% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_train_time.R
\name{get_final_ids}
\alias{get_final_ids}
\title{Helper function to get the final experiment IDs for the HC DataFrame}
\usage{
get_final_ids(df)
}
\arguments{
\item{df}{HC DataFrame}
}
\value{
DataFrame containing the final experiment IDs
}
\description{
Helper function to get the final experiment IDs for the HC DataFrame
}
\keyword{internal}
