% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_boot_distns.R
\name{get_plot_name}
\alias{get_plot_name}
\title{Helper function to get the plot name from the ... argument}
\usage{
get_plot_name(addl_args)
}
\arguments{
\item{addl_args}{Additional arguments passed by ...}
}
\value{
Plot name
}
\description{
Helper function to get the plot name from the ... argument
}
