% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tex_code.R
\name{get_uniq_ids}
\alias{get_uniq_ids}
\title{Gets the unique experiment IDs}
\usage{
get_uniq_ids(df)
}
\arguments{
\item{df}{DataFrame containing a column with the experiment pair IDs}
}
\value{
Vector containing the unique pair IDs
}
\description{
Gets the unique experiment IDs
}
