% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBootResults-package.R
\docType{package}
\name{getBootResults-package}
\alias{getBootResults}
\alias{getBootResults-package}
\title{getBootResults: Gets Bootstrapped Results}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Zachary Blanks \email{zachblanks17@gmail.com}

}
\keyword{internal}
