% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{infer_exp_combos}
\alias{infer_exp_combos}
\title{Infers if any of the experiments belong as a group for plotting purposes}
\usage{
infer_exp_combos(exp_diff_df, exp_df, pair_df)
}
\arguments{
\item{exp_diff_df}{Experiment difference DataFrame}

\item{exp_df}{DataFrame detailing all experiment settings}

\item{pair_df}{DataFrame defining the experiment pairs}
}
\value{
List containing the pair_ids that belong togeter and a DataFrame
used for plotting
}
\description{
Infers if any of the experiments belong as a group for plotting purposes
}
\keyword{internal}
