% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{infer_other_col}
\alias{infer_other_col}
\title{Helper function to determine the other distinguishing factor for a valid
combination}
\usage{
infer_other_col(pair_df, exp_df, pair_combo)
}
\arguments{
\item{pair_df}{DataFrame defining the experiment pairs}

\item{exp_df}{DataFrame detailing all experiment settings}

\item{pair_combo}{Combination of pair_ids that are valid}
}
\value{
DataFrame containing the pair_combo and the other distinguishing
column
}
\description{
Helper function to determine the other distinguishing factor for a valid
combination
}
\keyword{internal}
