% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{prep_boot_df}
\alias{prep_boot_df}
\title{Helper function to prepare the bootstrap DataFrame for plotting}
\usage{
prep_boot_df(boot_df, info_df)
}
\arguments{
\item{boot_df}{Bootstrap results DataFrame}

\item{info_df}{DataFrame containing the plotting info for the \code{boot_df}}
}
\value{
Updated DataFrame
}
\description{
Helper function to prepare the bootstrap DataFrame for plotting
}
\keyword{internal}
