% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_train_time.R
\name{update_method}
\alias{update_method}
\title{Updates the methods for a DataFrame}
\usage{
update_method(df, exp_df)
}
\arguments{
\item{df}{DataFrame to update the method column}

\item{exp_df}{Experiment settings DataFrame}
}
\value{
DataFrame with updated methods
}
\description{
Updates the methods for a DataFrame
}
\keyword{internal}
