% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bootstrap_plots.R
\name{gen_boot_plots}
\alias{gen_boot_plots}
\title{Generates all of the bootstrap distributions}
\usage{
gen_boot_plots(wd, max_combo_size = 2)
}
\arguments{
\item{wd}{Directory containing the files needed to generate the plots}

\item{max_combo_size}{Max combination size for experiment combos}
}
\value{
Nothing; saves the plots to disk
}
\description{
Generates all of the bootstrap distributions
}
