% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_train_time.R
\name{make_plot}
\alias{make_plot}
\title{Helper function to make the time comparison plot}
\usage{
make_plot(df, wd, metric, addl_args)
}
\arguments{
\item{df}{DataFrame containing the data needed to generate the plot}

\item{wd}{Location to save the plot}

\item{metric}{Which metric to use for the time comparison plot}

\item{addl_args}{Additional arguments passed by ...}
}
\value{
ggplot2 object
}
\description{
Helper function to make the time comparison plot
}
\keyword{internal}
